# docs/swagger.yaml
basePath: /api
definitions:
...
  RIP_internal_app_ds.ComponentCreateRequest:
    properties:
      description:
        type: string
      specifications:
        items:
          type: string
        type: array
      tdp:
        type: integer
      title:
        type: string
    required:
    - description
    - tdp
    - title
    type: object
  RIP_internal_app_ds.ComponentDTO:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      specifications:
        items:
          type: string
        type: array
      status:
        type: boolean
      tdp:
        type: integer
      title:
        type: string
    type: object
...
  RIP_internal_app_ds.ComponentUpdateRequest:
    properties:
      description:
        type: string
      specifications:
        items:
          type: string
        type: array
      tdp:
        type: integer
      title:
        type: string
    type: object
...
  RIP_internal_app_ds.PaginatedResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API-сервер для управления заявками и компонентами серверов в системе
    CoolRequest.
  title: API для системы FRAX
  version: "1.0"
paths:
...
  /components:
    get:
      description: Возвращает постраничный список компонентов системы охлаждения.
      parameters:
      - description: Фильтр по названию компонента
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RIP_internal_app_ds.PaginatedResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список компонентов (все)
      tags:
      - components
    post:
      consumes:
      - application/json
      description: Создает новую запись о компоненте системы охлаждения.
      parameters:
      - description: Данные нового компонента
        in: body
        name: componentData
        required: true
        schema:
          $ref: '#/definitions/RIP_internal_app_ds.ComponentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/RIP_internal_app_ds.ComponentDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (не модератор)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новый компонент (только модератор)
      tags:
      - components
  /components/{id}:
    delete:
      description: Удаляет компонент системы охлаждения из системы.
      parameters:
      - description: ID компонента для удаления
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить компонент (только модератор)
      tags:
      - components
    get:
      description: Возвращает детальную информацию о компоненте системы охлаждения.
      parameters:
      - description: ID компонента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RIP_internal_app_ds.ComponentDTO'
        "404":
          description: Компонент не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить один компонент по ID (все)
      tags:
      - components
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующем компоненте системы охлаждения.
      parameters:
      - description: ID компонента
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/RIP_internal_app_ds.ComponentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RIP_internal_app_ds.ComponentDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить компонент (только модератор)
      tags:
      - components
  /components/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает и привязывает изображение к компоненту системы охлаждения.
      parameters:
      - description: ID компонента
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: URL загруженного изображения
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Файл не предоставлен
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение для компонента (только модератор)
      tags:
      - components
...
  /coolrequests/draft/components/{component_id}:
    post:
      description: Находит или создает черновик заявки для текущего пользователя и
        добавляет в него компонент.
      parameters:
      - description: ID компонента для добавления
        in: path
        name: component_id
        required: true
        type: integer
      responses:
        "201":
          description: Сообщение об успехе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить компонент в черновик заявки (все)
      tags:
      - components
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"